Fix error message on bad imports
